Feat: Rename folder Models to DataModels, Add FinancialInformation class, Add mapping properties, Change relationship of JobApplicationHistory class with Status table
