+++ jobApplicationTracker/DataModels/Interview.cs
@@ -4,13 +4,13 @@ namespace jobApplicationTrackerApi.Models
 {
     public class Interview
     {
-        [Key]
         public Guid Id { get; set; }
         public Guid JobApplicationId { get; set; }
-        [Required]
-        [MaxLength(50, ErrorMessage = "Type should be less than 50 characters")]
         public string Type { get; set; }
         public DateTime? Date { get; set; }
         public string Notes { get; set; } = string.Empty;
+
+
+        public JobApplication JobApplication { get; set; }
     }
 }
