+++ jobApplicationTracker/DataModels/JobApplicationHistory.cs
@@ -4,10 +4,14 @@ namespace jobApplicationTrackerApi.Models
 {
     public class JobApplicationHistory
     {
-        [Key]
         public Guid Id { get; set; }
         public Guid JobApplicationId { get; set; }
-        public Guid StatusId { get; set; }
         public DateTime UpdatedAt { get; set; } = DateTime.Now;
+
+
+        public JobApplication JobApplication { get; set; }
+
+
+        public ICollection<JobApplication> JobApplications { get; set; }
     }
 }
